add aoc list changed event
